# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Creates a new user account",
        "operationId": "AuthController_register_api/v1",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique key used to safely retry the creation request. The same key cannot be reused with a different request payload.",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255,
              "example": "7b9f3e2a-5f1c-4f4b-9d8d-123456789abc"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Validation failed."
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Unauthorized",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You are not allowed to perform this action",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Resource not found",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "409": {
            "description": "User already exists."
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Failed",
                  "status": 412,
                  "detail": "Task has been modified. Please refresh and try again.",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Required",
                  "status": 428,
                  "detail": "If-Match header is required",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Too many requests",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "Internal server error",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticates a user and returns access and refresh tokens.",
        "operationId": "AuthController_login_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful."
          },
          "400": {
            "description": "Validation failed."
          },
          "401": {
            "description": "Invalid credentials."
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You are not allowed to perform this action",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Resource not found",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The request conflicts with the current state of the resource",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Failed",
                  "status": 412,
                  "detail": "Task has been modified. Please refresh and try again.",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Required",
                  "status": 428,
                  "detail": "If-Match header is required",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Too many requests",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "Internal server error",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          }
        },
        "summary": "Login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Generates a new access token using a valid refresh token.",
        "operationId": "AuthController_refresh_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token refreshed successfully."
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": "Validation failed",
                  "instance": "/api/v1/tasks",
                  "errors": [
                    "email must be an email",
                    "password must be longer than or equal to 8 characters"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token."
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You are not allowed to perform this action",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Resource not found",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The request conflicts with the current state of the resource",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Failed",
                  "status": 412,
                  "detail": "Task has been modified. Please refresh and try again.",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Required",
                  "status": 428,
                  "detail": "If-Match header is required",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Too many requests",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "Internal server error",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Logs out the authenticated user by invalidating the stored refresh token.",
        "operationId": "AuthController_logout_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully."
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Bad Request",
                  "status": 400,
                  "detail": "Validation failed",
                  "instance": "/api/v1/tasks",
                  "errors": [
                    "email must be an email",
                    "password must be longer than or equal to 8 characters"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You are not allowed to perform this action",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Resource not found",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The request conflicts with the current state of the resource",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Failed",
                  "status": 412,
                  "detail": "Task has been modified. Please refresh and try again.",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Required",
                  "status": 428,
                  "detail": "If-Match header is required",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Too many requests",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "Internal server error",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "description": "Verifies a user email address using a valid verification token.",
        "operationId": "AuthController_verifyEmail_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Email verified successfully."
          },
          "400": {
            "description": "Invalid or expired verification token."
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Unauthorized",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You are not allowed to perform this action",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Resource not found",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Conflict",
                  "status": 409,
                  "detail": "The request conflicts with the current state of the resource",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "412": {
            "description": "Precondition Failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Failed",
                  "status": 412,
                  "detail": "Task has been modified. Please refresh and try again.",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Precondition Required",
                  "status": 428,
                  "detail": "If-Match header is required",
                  "instance": "/api/v1/tasks/..."
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Too Many Requests",
                  "status": 429,
                  "detail": "Too many requests",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500,
                  "detail": "Internal server error",
                  "instance": "/api/v1/tasks"
                }
              }
            }
          }
        },
        "summary": "Verify email address",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Manager API",
    "description": "NestJS Task Manager API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank",
            "description": "URI reference identifying the problem type. Defaults to about:blank."
          },
          "title": {
            "type": "string",
            "example": "Bad Request",
            "description": "Short, human-readable summary of the problem."
          },
          "status": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code generated by the origin server."
          },
          "detail": {
            "type": "string",
            "example": "Validation failed",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/tasks",
            "description": "URI reference identifying the specific occurrence."
          },
          "requestId": {
            "type": "string",
            "example": "01JXYZ123456789",
            "description": "Optional correlation/request identifier."
          },
          "errors": {
            "example": [
              "email must be an email",
              "password must be longer than or equal to 8 characters"
            ],
            "description": "Optional validation or additional error details.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ]
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "Password123",
            "description": "User password (minimum 6 characters)",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "description": "Registered email address"
          },
          "password": {
            "type": "string",
            "example": "Password123",
            "description": "Account password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "Valid refresh token issued during login"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
...
    }
  }
}